fix: fix uninitialized members, narrowing casts, and Rule of Five violations
- profileform.cpp: add missing <memory> include for std::unique_ptr
- profileinfo.cpp: remove dead if/else with identical branches (V523)
- corefile.cpp: remove always-true guard after early-return avatar block (V547)
- message.h: initialise MessageMetadata::type to avoid UB (V730)
- history.h: initialise FileDbInsertionData::size and direction (V730)
- dbupgrader.cpp: use int64_t cast to avoid size_t->int narrowing (V220)
- dbto11.cpp: remove redundant clear() on moved-from vector (V1030)
- core.cpp: remove redundant Offline assignment already set at init (V1048)
- toxoptions.h/.cpp: add move assignment + explicit delete of copies (V690)

+++ src/persistence/db/upgrades/dbupgrader.cpp
@@ -251,10 +251,10 @@ bool DbUpgrader::dbSchemaUpgrade(std::shared_ptr<RawDatabase>& db, IMessageBoxMa
                                                  dbSchema6to7,  dbSchema7to8,          dbSchema8to9,
                                                  dbSchema9to10, DbTo11::dbSchema10to11};
 
-    assert(databaseSchemaVersion < static_cast<int>(upgradeFns.size()));
+    assert(databaseSchemaVersion < static_cast<int64_t>(upgradeFns.size()));
     assert(upgradeFns.size() == SCHEMA_VERSION);
 
-    for (int64_t i = databaseSchemaVersion; i < static_cast<int>(upgradeFns.size()); ++i) {
+    for (int64_t i = databaseSchemaVersion; i < static_cast<int64_t>(upgradeFns.size()); ++i) {
         const auto newDbVersion = i + 1;
         if (!upgradeFns[i](*db)) {
             qCritical() << "Failed to upgrade db to schema version" << newDbVersion << " aborting";
